Move generic kernels to opt_kernel
