+++ src/ngraph/pass/constant_folding_broadcast.cpp
@@ -16,7 +16,7 @@
 
 #include "constant_folding.hpp"
 #include "ngraph/op/broadcast.hpp"
-#include "ngraph/runtime/generic_cpu/kernel/broadcast.hpp"
+#include "ngraph/runtime/opt_kernel/broadcast.hpp"
 #include "ngraph/type/element_type.hpp"
 
 using namespace std;
@@ -45,11 +45,11 @@ shared_ptr<op::Constant> fold_constant_broadcast(shared_ptr<op::Constant> consta
         auto static_bcast_axes = broadcast_v1->get_broadcast_axes();
         if (static_bcast_axes.first)
         {
-            runtime::gcpu::kernel::broadcast<T>(constant->get_data_ptr<T>(),
-                                                data_ptr,
-                                                constant->get_shape(),
-                                                out_shape,
-                                                static_bcast_axes.second);
+            runtime::opt_kernel::broadcast<T>(constant->get_data_ptr<T>(),
+                                              data_ptr,
+                                              constant->get_shape(),
+                                              out_shape,
+                                              static_bcast_axes.second);
         }
         else
         {
@@ -58,11 +58,11 @@ shared_ptr<op::Constant> fold_constant_broadcast(shared_ptr<op::Constant> consta
     }
     else if (auto broadcast_v0 = as_type_ptr<op::v0::Broadcast>(broadcast))
     {
-        runtime::gcpu::kernel::broadcast<T>(constant->get_data_ptr<T>(),
-                                            data_ptr,
-                                            constant->get_shape(),
-                                            out_shape,
-                                            broadcast_v0->get_broadcast_axes());
+        runtime::opt_kernel::broadcast<T>(constant->get_data_ptr<T>(),
+                                          data_ptr,
+                                          constant->get_shape(),
+                                          out_shape,
+                                          broadcast_v0->get_broadcast_axes());
     }
     else
     {
